Inline integrateMartech (#260)
